backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
drf-yasg==1.21.10
pandas==2.3.1
pillow==11.3.0
django-cors-headers==4.7.0
asgiref==3.8.1
sqlparse==0.3.1
djangorestframework-simplejwt==5.3.0